try moving depenencies to a hidden folder
